pyproject.toml
[tool.poetry]
name = "rag-demo"
version = "0.1.0"
description = ""
authors = ["Andrea Stefanachi <andreai@stefanachi.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.1"
unstructured = "^0.11.2"
langchain = "^0.0.344"
python-dotenv = "^1.0.0"
chromadb = "^0.4.18"
pypdf = "^3.17.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
